Moved Flow-related classes and files to opm/simulators/flow.
