+++ opm/simulators/flow/BlackoilModelEbos.hpp
@@ -29,8 +29,8 @@
 
 #include <opm/simulators/timestepping/AdaptiveTimeSteppingEbos.hpp>
 
-#include <opm/autodiff/NonlinearSolverEbos.hpp>
-#include <opm/autodiff/BlackoilModelParametersEbos.hpp>
+#include <opm/simulators/flow/NonlinearSolverEbos.hpp>
+#include <opm/simulators/flow/BlackoilModelParametersEbos.hpp>
 #include <opm/simulators/wells/BlackoilWellModel.hpp>
 #include <opm/simulators/aquifers/BlackoilAquiferModel.hpp>
 #include <opm/simulators/wells/WellConnectionAuxiliaryModule.hpp>
