+++ opm/simulators/wells/StandardWell.hpp
@@ -24,9 +24,9 @@
 #define OPM_STANDARDWELL_HEADER_INCLUDED
 
 
+#include <opm/simulators/wells/RateConverter.hpp>
 #include <opm/simulators/wells/WellInterface.hpp>
 #include <opm/simulators/linalg/ISTLSolverEbos.hpp>
-#include <opm/autodiff/RateConverter.hpp>
 
 #include <opm/material/densead/DynamicEvaluation.hpp>
